[fix] Make `qd.clz` work on u32 / u64 across all backends
`qd.clz(u32_value)` was emitting QD_NOT_IMPLEMENTED on CUDA and produced
undefined results on SPIR-V (GLSL.std.450 FindSMsb is undefined for the
all-bits-set case).  The new `subgroup.segmented_reduce_add` is the
first user of `clz` in the codebase and exposed both bugs.

* CUDA: route u32 / u64 inputs through the same `__nv_clz` / `__nv_clzll`
  intrinsics used for i32 / i64 — the underlying bit pattern is what
  matters, the C declaration on signed types is a header-level
  convention.
* SPIR-V: dispatch to FindUMsb (#75) for unsigned inputs and FindSMsb
  (#74) for signed.  The two GLSL.std.450 instructions return a value
  of the same type as their operand, so add an explicit OpBitcast back
  to i32 before the `32 - msb - 1` subtraction (otherwise SPIR-V's
  strict-type `sub` asserts on mixed i32 / u32).
* Python: in `segmented_reduce_add`, wrap `clz`'s result in `i32(...)`
  so the subsequent arithmetic is uniformly signed-32-bit (the trace-
  time tracer would otherwise propagate u32 from the input through to
  the subtraction, hitting SPIR-V's same-type assertion).

Tests: `subgroup.segmented_reduce_add` tests now pass on CUDA + Vulkan
across i32 / i64 / u64 / f32 / f64 and `log2_size` in [0..5], including
the all-heads, no-heads, and truthy-predicate edge cases.

+++ python/quadrants/lang/simt/subgroup.py
@@ -312,7 +312,10 @@ def segmented_reduce_add(value, head_flag, log2_size: template()):
     effective_mask = (head_mask & group_mask) | (u32(1) << group_base)
     inclusive_mask = u32(0xFFFFFFFF) >> u32(i32(31) - lane)
     lower = effective_mask & inclusive_mask
-    segment_head = i32(31) - clz(lower)
+    # `clz` follows the input type on most backends (CUDA explicitly normalizes to i32; AMDGPU LLVM `ctlz` returns
+    # input type; SPIR-V FindUMsb returns the input type).  Wrap the result in `i32(...)` so the subsequent arithmetic
+    # is uniformly signed-32-bit and SPIR-V's strict-type `sub` is happy.
+    segment_head = i32(31) - i32(clz(lower))
     distance = lane - segment_head
     for i in impl.static(range(log2_size)):
         offset = impl.static(1 << i)
